move verifier under pkg/agents
