+++ pkg/agents/verifier/__init__.py
@@ -1,0 +1,1 @@
+# Package for verifier agents
